requirements.txt
# Core Deep Learning Framework
torch>=2.0.0
torchvision>=0.15.0

# Web Application Framework
streamlit>=1.28.0

# Computer Vision and Image Processing
opencv-python>=4.8.0
Pillow>=10.0.0

# Scientific Computing
numpy>=1.24.0
scikit-learn>=1.3.0

# Data Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
tqdm>=4.65.0

# Optional: For CUDA support (uncomment if using GPU)
# torch-audio>=2.0.0

# Development Dependencies (optional)
# jupyter>=1.0.0
# ipykernel>=6.25.0